fix two more modals & PostPage into ts
