+++ web/src/pages/Post/Post.css
@@ -13,29 +13,8 @@
     width: 100%;
 }
 
-/* The Modal (background) */
-.post-detail-new-post-modal {
-    /*display: none; !* Hidden by default *!*/
-    position: fixed; /* Stay in place */
-    z-index: 1; /* Sit on top */
-    left: 0;
-    top: 0;
-    width: 100%; /* Full width */
-    height: 100%; /* Full height */
-    overflow: auto; /* Enable scroll if needed */
-    background-color: rgb(0,0,0); /* Fallback color */
-    background-color: rgba(0,0,0,0.4); /* Black w/ opacity */
-}
-
-/* Modal Content/Box */
-.post-detail-new-post-modal-content{
-    margin: 15% auto; /* 15% from the top and centered */
-    width: 80%; /* Could be more or less, depending on screen size */
-}
-
 @media only screen and (max-width: 750px) {
     .post-wrapper-page {
         padding: 0 0 50px 0;
     }
-
 }
